refactored word count functionality to its own class, added unit tests

+++ src/main/java/WordCounter.java
@@ -1,3 +1,3 @@
 public interface WordCounter {
-    Integer countWords(String line);
+    long countWords(String line);
 }
